Criação das Validações Básicas dos CRUDs

+++ api/src/core/application/use-cases/custumer/commands/delete-customer.use-command.ts
@@ -1,18 +1,30 @@
 import { CustomerRepository } from "src/core/infrastructure/Repositories/customer.repository"
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs"
+import { Result, result } from "src/core/infrastructure/Shared/result.util";
+import { messages } from "src/core/infrastructure/Shared/messages";
+import { BadRequestException, NotFoundException } from "@nestjs/common";
 
 export class DeleteCustomerCommand {
   constructor(public readonly id: number) {}
 }
 
 @CommandHandler(DeleteCustomerCommand)
-export class DeleteCustomerHandler implements ICommandHandler<DeleteCustomerCommand, { id: number }> {
+export class DeleteCustomerHandler implements ICommandHandler<DeleteCustomerCommand, Result<{ id: number }>> {
 
   constructor(private readonly repository: CustomerRepository) {
   }
 
-  async execute(command: DeleteCustomerCommand): Promise<{ id: number }> {
+  async execute(command: DeleteCustomerCommand): Promise<Result<{ id: number }>> {
+
+
+    const registerExists = await this.repository.getFirstByParameters({
+      id: command.id,
+    });
+
+    if (!registerExists)
+      throw new NotFoundException(messages.CUSTOMER_NOT_FOUND(command.id));
+
     const deletedId = await this.repository.delete(command.id);
-    return { id: deletedId };
+    return result({ id: deletedId }).Success();
   }
 }
